requirements.txt
certifi==2021.5.30
click==8.0.1
cycler==0.10.0
Cython==0.29.24
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
opencv-python-headless==4.5.3.56
pandas==1.3.2
Pillow==8.3.1
pip==21.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
setuptools==52.0.0
six==1.16.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wheel==0.37.0
zipp==3.5.0
